Fixes rails completion
Currently, the rails completion spec checks if rails is in the
Gemfile, if there isn't a gemfile or if rails isn't in it the only
completion for rails will be `new`.

The problem, is that it checks for rails using `\grep "gem 'rails'"`.
The problem with this is that it only matches when 'single-quotes'
are used, if "double-quotes" are used, grep won't match anything and
only `new` will be suggested.

This commit makes it match double and single quotes using
`\grep "gem ['\"]rails['\"]"`.