Add glow
I wanted to add the 'wrapwidth' option to neovim, but to do that I had
to read the documentation. To read the documentation, I wanted to use
glow. In the documentation was a link. I wanted to see if there was a
flag to use the proper escape codes to make the link clickable, because
it was wrapping over two lines. There were no fig completions for glow.
Well, I _was_ going to add an option to neovim, but I guess I'm just
contributing another fig spec lol