feat: add lsof spec
In this commit, we want to add the lsof spec.

I basically cross referenced the `lsof -h` command with `man lsof` and
did my best to make it easy to understand most of the args.

Some of the args I had to do some word-smithing to make it short
enough to fit inside the autocomplete window.

I tried to make the -i option more dynamic, and added fancy ways
of grabbing ip address or service, etc...